Consolidated and updated JavaScript
