+++ include/footer.php
@@ -26,6 +26,5 @@ if (__FILE__ === $_SERVER['SCRIPT_FILENAME']) {
     </footer>
     <!-- JS -->
     <script type="module" src="assets/js/tts.js?v=<?php echo $tts::version; ?>"></script>
-<?php if (in_array(basename($_SERVER['SCRIPT_FILENAME']), ['conversation.php', 'playlist.php'])) echo '    <script type="module" src="assets/js/conversations.js?v=' . $tts::version . '"></script>' . PHP_EOL; ?>
     </body>
 </html>
